Hoist calculate_components to RouteCalculation
